refactor: extract clean `graph` CLI group into cli_graph.py
The `graph` command group (compare, explain) lived inside cli_buildsource.py
— the `collect` module — and its two graph-only helpers sat among the collect
helpers, though nothing in `collect` used them. Make it clean and logical:

- New sibling module `cli_graph.py` holds the group, both commands, and the
  co-located `_load_source_graph` / `_resolve_symbol_from_report` helpers,
  per the "command group → cli_<name>.py" convention. Registered via the
  side-effect import block in cli.py.
- De-duplicate the `--format text|json` option into one shared `_format_option`
  decorator; this also fixes `graph explain --format` which had no help text.
- Add an intentional `* explain` panel to cli_help.OPTION_GROUPS so its inputs
  group ahead of the output flag, instead of relying on the accidental
  `* compare` fnmatch collision (documented in a comment).
- Replace repeated `import json as _json` with a single module-level import.
- Wire up mypy override, import-cycle allowlist, and the test import.

No behaviour change: same commands, args, output. cli_buildsource.py drops
~165 lines and off the large-file warning list.
