+++ abicheck/cli_buildsource.py
@@ -62,7 +62,6 @@ from .cli_buildsource_helpers import (  # noqa: F401  (re-exported for API stabi
     _intrinsic_coverage as _intrinsic_coverage,
     _layer_presence as _layer_presence,
     _load_pack_or_raise as _load_pack_or_raise,
-    _load_source_graph as _load_source_graph,
     _merge_attach_combined as _merge_attach_combined,
     _merge_fold_packs as _merge_fold_packs,
     _merge_handle_conflicts as _merge_handle_conflicts,
@@ -72,7 +71,6 @@ from .cli_buildsource_helpers import (  # noqa: F401  (re-exported for API stabi
     _optional_coverage as _optional_coverage,
     _purge_external_outputs as _purge_external_outputs,
     _resolve_side_pack as _resolve_side_pack,
-    _resolve_symbol_from_report as _resolve_symbol_from_report,
     _run_adapters as _run_adapters,
     _run_external_extractors as _run_external_extractors,
     attach_evidence_metrics as attach_evidence_metrics,
@@ -1035,168 +1033,3 @@ def merge_cmd(
 
     output.write_text(snapshot_to_json(base), encoding="utf-8")
     _merge_print_summary(base_path, contributors, len(snaps), combined, output)
-
-
-@main.group("graph")
-def graph_group() -> None:
-    """L5 source-graph tools (ADR-031): structural diff and finding localization.
-
-    These commands *explain and prioritize* impact through the source→symbol
-    graph; per ADR-028 D3 they never, on their own, decide or suppress an
-    artifact-proven ABI break.
-    """
-
-
-@graph_group.command("compare")
-@click.argument("old", type=click.Path(path_type=Path))
-@click.argument("new", type=click.Path(path_type=Path))
-@click.option(
-    "--format",
-    "fmt",
-    default="text",
-    show_default=True,
-    type=click.Choice(["text", "json"], case_sensitive=False),
-    help="Output format for the structural graph diff.",
-)
-def compare_graph_cmd(old: Path, new: Path, fmt: str) -> None:
-    """Compare two L5 source graph summaries (ADR-031 D6, D8).
-
-    \b
-    OLD and NEW may each be a `graph/source_graph_summary.json` file or an
-    evidence-pack directory produced by `collect --source-graph summary`.
-
-    The diff is structural — which nodes/edges entered or left the graph. Per
-    ADR-028 D3 / ADR-031 D6 it *explains and prioritizes* impact; it never, on
-    its own, decides or suppresses an artifact-proven ABI break.
-    """
-    import json as _json
-
-    from .buildsource.source_graph import diff_source_graph, diff_source_graph_findings
-
-    old_graph = _load_source_graph(old)
-    new_graph = _load_source_graph(new)
-    delta = diff_source_graph(old_graph, new_graph)
-    findings = diff_source_graph_findings(old_graph, new_graph)
-
-    if fmt == "json":
-        payload = delta.to_dict()
-        payload["findings"] = [
-            {
-                "kind": c.kind.value,
-                "symbol": c.symbol,
-                "description": c.description,
-                "old_value": c.old_value,
-                "new_value": c.new_value,
-            }
-            for c in findings
-        ]
-        click.echo(_json.dumps(payload, indent=2, sort_keys=True))
-        return
-
-    if not delta.changed:
-        click.echo("Source graphs are structurally identical.")
-        click.echo(f"  graph_id: {old_graph.graph_id or old_graph.compute_graph_id()}")
-        return
-
-    click.echo("Source graph structural diff:")
-    click.echo(
-        f"  nodes: +{len(delta.added_nodes)} / -{len(delta.removed_nodes)}    "
-        f"edges: +{len(delta.added_edges)} / -{len(delta.removed_edges)}"
-    )
-    for node in delta.added_nodes:
-        click.echo(f"  + node [{node.kind}] {node.label or node.id}")
-    for node in delta.removed_nodes:
-        click.echo(f"  - node [{node.kind}] {node.label or node.id}")
-    for edge in delta.added_edges:
-        click.echo(f"  + edge {edge.kind}: {edge.src} -> {edge.dst}")
-    for edge in delta.removed_edges:
-        click.echo(f"  - edge {edge.kind}: {edge.src} -> {edge.dst}")
-
-    if findings:
-        # Graph-derived RISK findings (ADR-031 D6): explanation/prioritization,
-        # never a standalone ABI-break verdict (ADR-028 D3).
-        click.echo(f"\nGraph-derived risk findings ({len(findings)}):")
-        for c in findings:
-            click.echo(f"  [{c.kind.value}] {c.symbol}: {c.description}")
-
-
-@graph_group.command("explain")
-@click.option(
-    "--sources",
-    "sources",
-    type=click.Path(path_type=Path),
-    required=True,
-    help="Source/graph pack directory (or a source_graph_summary.json) to explain through.",
-)
-@click.option(
-    "--symbol",
-    "symbol",
-    default="",
-    help="Exported (mangled) binary symbol to localize.",
-)
-@click.option(
-    "--report",
-    "report",
-    type=click.Path(path_type=Path),
-    default=None,
-    help="A `compare --format json` report; with --finding-id, resolves the symbol from it.",
-)
-@click.option(
-    "--finding-id",
-    "finding_id",
-    default="",
-    help="Index (or symbol) of a finding in --report to localize.",
-)
-@click.option(
-    "--format",
-    "fmt",
-    default="text",
-    show_default=True,
-    type=click.Choice(["text", "json"], case_sensitive=False),
-)
-def explain_finding_cmd(
-    sources: Path,
-    symbol: str,
-    report: Path | None,
-    finding_id: str,
-    fmt: str,
-) -> None:
-    """Localize a finding through L5 source-graph evidence (ADR-031 D8).
-
-    Given an exported symbol (directly via --symbol, or resolved from a
-    `--report` finding via --finding-id), walks the graph to show what produced
-    and reaches it: exporting target, source declaration(s), declaring public
-    header(s), ABI-relevant build option(s), and static callees. This explains
-    and prioritizes; it is never an ABI verdict (ADR-031 D6).
-    """
-    import json as _json
-
-    from .buildsource.source_graph import localize_symbol
-
-    graph = _load_source_graph(sources)
-    if not symbol and report is not None:
-        symbol = _resolve_symbol_from_report(report, finding_id)
-    if not symbol:
-        raise click.ClickException(
-            "No symbol to explain: pass --symbol, or --report with --finding-id."
-        )
-
-    result = localize_symbol(graph, symbol)
-    if fmt == "json":
-        click.echo(_json.dumps(result, indent=2, sort_keys=True))
-        return
-
-    click.echo(f"Explaining symbol: {symbol}")
-    if not result["found"]:
-        click.echo(
-            "  (symbol not present in the source graph — no localization available)"
-        )
-    rows = [
-        ("exported by target(s)", result["exported_by_targets"]),
-        ("source declaration(s)", result["source_declarations"]),
-        ("declared in header(s)", result["declared_in_headers"]),
-        ("reached by build option(s)", result["reached_by_build_options"]),
-        ("static callee(s)", result["static_callees"]),
-    ]
-    for label, values in rows:
-        click.echo(f"  {label}: {', '.join(values) if values else '(none in graph)'}")
