+++ abicheck/cli_buildsource_helpers.py
@@ -557,82 +557,6 @@ def _echo_capabilities(
             click.echo(f"  [off] {label} — {why_off}", err=True)
 
 
-# ── graph compare: structural graph-to-graph diff (ADR-031 D6, D8) ────────────
-
-
-def _load_source_graph(path: Path) -> SourceGraphSummary:
-    """Load a source graph summary from a JSON file or an evidence-pack dir.
-
-    Accepts either ``…/graph/source_graph_summary.json`` directly or a pack
-    directory (the graph is read from its manifest layout). Raises a Click error
-    when neither yields a graph so the failure is actionable.
-    """
-    import json as _json
-
-    from .buildsource.source_graph import SourceGraphSummary
-
-    if path.is_dir():
-        pack = _load_pack_or_raise(path)
-        if pack.source_graph is None:
-            raise click.ClickException(
-                f"Evidence pack at {path} has no L5 source graph "
-                "(collect it with `collect --source-graph summary`)."
-            )
-        return pack.source_graph
-    if not path.is_file():
-        raise click.ClickException(f"No source graph summary at {path}.")
-    try:
-        data = _json.loads(path.read_text(encoding="utf-8"))
-    except (OSError, ValueError) as exc:
-        raise click.ClickException(
-            f"Cannot read source graph at {path}: {exc}"
-        ) from exc
-    if not isinstance(data, dict):
-        raise click.ClickException(f"{path} must contain a JSON object.")
-    # SourceGraphSummary.from_dict is intentionally forgiving (it defaults a
-    # missing nodes/edges to empty), so guard here: an unrelated JSON file (e.g.
-    # a pack manifest) would otherwise load as an empty graph and report a bogus
-    # diff instead of an actionable error.
-    if not isinstance(data.get("nodes"), list) or not isinstance(
-        data.get("edges"), list
-    ):
-        raise click.ClickException(
-            f"{path} is not a source graph summary "
-            "(expected top-level 'nodes' and 'edges' lists)."
-        )
-    return SourceGraphSummary.from_dict(data)
-
-
-def _resolve_symbol_from_report(report: Path, finding_id: str) -> str:
-    """Resolve a symbol from a `compare --format json` report finding.
-
-    ``finding_id`` may be a 0-based index into the report's changes, or a symbol
-    substring to match. Returns the matched change's ``symbol`` (or "").
-    """
-    import json as _json
-
-    try:
-        data = _json.loads(Path(report).read_text(encoding="utf-8"))
-    except (OSError, ValueError) as exc:
-        raise click.ClickException(f"Cannot read report {report}: {exc}") from exc
-    changes = data.get("changes") or data.get("findings") or []
-    if not isinstance(changes, list):
-        return ""
-    if finding_id.isdigit():
-        idx = int(finding_id)
-        if 0 <= idx < len(changes) and isinstance(changes[idx], dict):
-            return str(changes[idx].get("symbol", ""))
-        return ""
-    for change in changes:
-        if (
-            isinstance(change, dict)
-            and finding_id
-            and finding_id in str(change.get("symbol", ""))
-        ):
-            return str(change.get("symbol", ""))
-    return ""
-
-
 def _build_coverage(
     merged: BuildEvidence,
     has_build: bool,
