+++ pyproject.toml
@@ -117,6 +117,7 @@ module = [
     "abicheck.cli_debian_symbols",
     "abicheck.cli_baseline",
     "abicheck.cli_buildsource",
+    "abicheck.cli_graph",
     "abicheck.cli_plugin",
     "abicheck.cli_pr_comment",
     "abicheck.cli_probe",
